example/requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
blis==0.7.11
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.15.1
cmake==3.27.7
confection==0.1.3
cymem==2.0.8
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.7.0/de_core_news_sm-3.7.0-py3-none-any.whl#sha256=d88c737eb7eb766f730f6a2dcb99dfcdb81623e1e0d89a9c638a2182ac19c52e
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
filelock==3.12.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
langcodes==3.3.0
lit==17.0.2
logic11-cpp==0.0.0
logic11-cuda==0.0.0
Markdown==3.5
MarkupSafe==2.1.3
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
packaging==23.2
pathy==0.10.2
Pillow==10.0.1
pipreqs==0.4.13
portalocker==2.8.2
preshed==3.0.9
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
smart-open==6.4.0
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
thinc==8.2.1
torchdata==0.6.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.0.6
wasabi==1.1.2
weasel==0.3.2
Werkzeug==3.0.0
yarg==0.1.9
zipp==3.17.0
